Update to new OnMasteryStatChanged structure; Split Assassination spells/talents to individual files.
